+++ util/Common.py
@@ -1,4 +1,10 @@
-from ipaddress import ip_address, ip_network, IPv4Address
+"""
+@Copyrights 2020 Rakesh Kumar T
+github: rakesht2499
+"""
+
+from ipaddress import ip_address, IPv4Address
+
 
 def to_number(ip):
     if isinstance(ip_address(ip), IPv4Address):
@@ -6,15 +12,3 @@ def to_number(ip):
     else:
         padding_len = 39
     return str(int(ip_address(ip))).rjust(padding_len, "0")
-
-
-class Common:
-    def _insert_cidr(self, cidr) -> bool:
-        ins = True
-        for ip in ip_network(cidr):
-            ins = ins and self.add(str(ip))
-        return ins
-
-    def _remove_cidr(self, cidr) -> bool:
-        for ip in ip_network(cidr):
-            self.remove(str(ip))